requirements.txt
playwright
python-dotenv==1.0.1
pandas
boto3
